tests/requirements-test.txt
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0
httpx==0.25.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch==2.1.0
torchvision==0.16.0
Pillow==10.0.1
numpy==1.24.3
pydantic==2.4.2
python-multipart==0.0.6
Jinja2==3.1.2
aiofiles==23.2.1
seaborn==0.13.0
scikit-learn==1.3.0
pandas==2.1.0
matplotlib==3.7.0
opencv-python==4.8.1.78
albumentations==1.3.1
sqlite3